nvmeof-recoverer: modify OSD activation process to store metadata on OSD device
This commit updates the OSD deployment specification to store metadata directly on the OSD device partition instead of the host server.
- Modified the OSD activation process to mount the OSD device partition before starting each container in the OSD pod.
- Added regex to extract the OSD device path when the OSD device is given as a partition.

These changes leverage the fault domain of NVMe-oF devices, allowing failed OSDs to be revived on another node using just the device mount, without the need for metadata backups.