nvmeof-recoverer: watch for osd pod fault events
this commit modifies the controller to watch for events where an osd pod transitions to a failed state or is marked for deletion.
- added predicates to filter for osd pod death events in the watch function
- updated reconcile logic to handle nvmeofstorage and rook-ceph-osd requests differently
- added helper functions isOSDPod and isPodDead to detect relevant pod states

the logic for reconciling rook-ceph-osd requests will be implemented in a future commit to handle updating node attachment information in the nvmeofstorage cr when an osd pod dies.

Signed-off-by: cheolho.kang <cheolho.kang@samsung.com>