nvmeof-recoverer: update nvmeofstorage CR with assigned osdID during CRUSH map update
this commit enhances the nvmeofstorage controller to update the nvmeofstorage CR with the osdID assigned to each device when modifying the CRUSH map.
this ensures that each device's assigned osdID is recorded in the nvmeofstorage CR.
- modified types.go to add 'osdID' to the FabricDevice struct
- modified UpdateCrushMapForOSD to return osdID
- added updateCR function to modify the nvmeofstorage CR
- updated reconcile function to save osdID in the nvmeofstorage CR after updating the CRUSH map
- 'deploy/' and 'Documentation/' directories are auto-generated code

these changes ensure that the nvmeofstorage CR accurately reflects the osd assignments, which is crucial for handling device faults and reassignments.

signed-off-by: cheolho.kang <cheolho.kang@samsung.com>